Improve controls for read manga and add possible read in list
